+++ src/Milestones/Block.php
@@ -19,11 +19,11 @@ class Block {
 				'attributes'      => [
 					'title'       => [
 						'type'    => 'string',
-						'default' => __( 'Back to School Fundraiser', 'give' ),
+						'default' => __( 'We\'ve raised {total} so far!', 'give' ),
 					],
 					'description' => [
 						'type'    => 'string',
-						'default' => __( 'This is a sample description.', 'give' ),
+						'default' => __( 'But we still need {total_remaining} to reach our goal!', 'give' ),
 					],
 					'image'       => [
 						'type'    => 'string',